Test for Task Status, Project, User Type get all

+++ oci-react-samples/MtdrSpring/backend/src/main/java/com/springboot/MyTodoList/repository/ProjectRepository.java
@@ -2,6 +2,8 @@ package com.springboot.MyTodoList.repository;
 
 import com.springboot.MyTodoList.model.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
@@ -11,7 +13,6 @@ import javax.transaction.Transactional;
 @Transactional
 @EnableTransactionManagement
 
-public interface ProjectRepository extends JpaRepository<Project,Integer> {
-
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 
 }
